switch to openai API
