+++ src/lib/utils.ts
@@ -1,7 +1,6 @@
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
 import { PrismaClient } from "@prisma/client";
-import { BigQueryPlatform } from "@/dataplatforms/BigQuery";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
